consolidate telemetry.rs
