+++ bonsai/src/status.rs
@@ -10,6 +10,7 @@ use serde::{Deserialize, Serialize};
 /// * Running, if the action is asynchronous and it needs more time to complete
 #[derive(Copy, Clone, PartialEq, Eq, Debug)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[cfg_attr(feature = "serde", serde(rename_all = "lowercase"))]
 pub enum Status {
     /// The behavior or action succeeded.
     Success,
